+++ app/src/main/java/com/czf/aviplayer/VideoActivity.java
@@ -93,7 +93,7 @@ public class VideoActivity extends AppCompatActivity {
   }
 
   private void renderFrame(Surface surface, Bitmap bp) {
-    int result = NativeLibInterface.setFrame(avi, bp);
+    long result = NativeLibInterface.setFrame(avi, bp);
     if (result != -1) {
       Canvas canvas = surface.lockCanvas(null);
       canvas.drawBitmap(bp, 0, 0, null);
